fix: address all code review feedback from opus
BLOCKERS:
- Add NavigationDrawerButton to ProwlarrTab (matches other tabs)
- iOS: Fix ForEach key to use \.offset (guid is nullable)
- iOS: Fix deprecated onChange to two-parameter form
- Android: Fix grabTarget null timing — use grabbingGuid state so
  per-card spinner actually shows during grab operation
- iOS: Add grabbingGuid for per-item grab state tracking

PATTERNS:
- Rename IndexersState → ProwlarrIndexersState (namespaced)
- Repository: encapsulate prowlarrClient, expose getIndexers/search
- Use cases: call repository methods instead of client directly

API MODELS:
- ProwlarrIndexer.message: String? → ProwlarrIndexerMessage?
- ProwlarrSearchResult: add Boolean defaults (downloadAllowed, fullSeason)
- ProwlarrGrabPayload.indexerId: Long → Int (API expects int)

STRINGS:
- Add MR strings: no_indexers_configured, loading_indexers,
  prowlarr_search_hint, search_releases_placeholder, enabled, disabled
- Android: Replace all hardcoded strings with mokoString()
- iOS: Replace all hardcoded strings with MR.strings().xxx.localized()

UI POLISH:
- Android: Add background color to indexer status dot
- Android: Use Box instead of Spacer for status indicator
- iOS: Add protocolDisplayName() helper (clean names, not enum desc)
- iOS: Use \.offset key in ForEach (avoid nullable guid crash)

+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/arr/api/client/ProwlarrClient.kt
@@ -50,6 +50,6 @@ class ProwlarrClient(
     suspend fun grab(guid: String, indexerId: Long): NetworkResult<ProwlarrSearchResult> =
         httpClient.safePost("$baseUrl/release") {
             contentType(ContentType.Application.Json)
-            setBody(ProwlarrGrabPayload(guid = guid, indexerId = indexerId))
+            setBody(ProwlarrGrabPayload(guid = guid, indexerId = indexerId.toInt()))
         }
 }
