+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/arr/api/model/ProwlarrGrabPayload.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ProwlarrGrabPayload(
     val guid: String,
-    val indexerId: Long
+    val indexerId: Int
 )
